fixed some pvs warnings

+++ Userland/SampleCodeModule/include/user_syscalls.h
@@ -77,7 +77,7 @@ int usys_total_space();
 
 // semaphores syscalls
 
-uint8_t usys_sem_open(char * name, int value);
+int16_t usys_sem_open(char * name, int value);
 
 void usys_sem_close(char * name);
